Application layers Dockerfile with JarLauncher demo
This commit improves our Dockerfile to reproduce the structure that the
JarLauncher expects. That way the classpath is going to be built
automatically and the main class detected from the manifest.

Note that Spring Boot has a CLASSPATH.IDX that is used to order the
classpath in a consistent manner.

From the root directory of our project, we can build this image provided
that the Docker daemon is available:

docker build -t scribe-custom:0.0.1-SNAPSHOT .

We can run the app in foreground as follows:

docker run --rm -p 8080:8080 scribe-custom:0.0.1-SNAPSHOT

If we want to set an environment variable to customize the application's
configuration, we can use the `-e` flag:

docker run --rm -p 8080:8080 -e CUSTOM_TEST_PROPERTY=cmdline\
    scribe-custom:0.0.1-SNAPSHOT

We can finally inspect the generated image with `dive` to confirm the
application is built out of a single layer:

dive scribe-custom:0.0.1-SNAPSHOT