Application layers Dockerfile with Spring Boot demo
Spring Boot enables layers mode by default. When this mode is enabled,
a small utility is included that lets us see the layers and manage them.

For instance, this command lists the available layers:

java -Djarmode=layertools -jar scribe/target/scribe-*.jar list

Out-of-the-box, four layers are defined:

* dependencies
* spring-boot-loader
* snapshot-dependencies
* application

This commit updates our Dockerfile to run the layertools to extract the
layers rather than doing this manually. This has the advantage of
creating an optimized structure from the get-go that simplifies the
copy instructions.

In particular, the classpath.idx file is stored in the application
layer so that it doesn't require an explicit copy instruction.

From the root directory of our project, we can build this image provided
that the Docker daemon is available:

docker build -t scribe-custom:0.0.1-SNAPSHOT .

We can run the app in foreground as follows:

docker run --rm -p 8080:8080 scribe-custom:0.0.1-SNAPSHOT

If we want to set an environment variable to customize the application's
configuration, we can use the `-e` flag:

docker run --rm -p 8080:8080 -e CUSTOM_TEST_PROPERTY=cmdline\
    scribe-custom:0.0.1-SNAPSHOT

We can finally inspect the generated image with `dive` to confirm the
application is built out of a single layer:

dive scribe-custom:0.0.1-SNAPSHOT