Graceful Shutdown demo
Spring Boot 2.3 introduces a graceful shutdown feature that provides a
grace period for existing request when application shutdown is
initiated.

Our generator adds a delay when it has to render a text that contains
"delay". We can use this and then shutdown the renderer. If we do so,
we can see that the circuit breaker was invoked as the app was
terminated and prevented the request to be processed.

If we add "server.shutdown=graceful" to the converter's configuratin, we
can see that the application initiates a graceful shutdown and lets
requests to complete before terminating. The default grace period is 30
seconds and can be tuned.

On Kubernetes, when the application is shutting down, some asynchronous
process may not have completed and traffic might still be routed to the
application. To avoid that, it is recommended to setup a `preStop` hook
that waits a bit before invoking the shutdown.